pyproject.toml
[project]
name = "song-recommender"
version = "0.1.0"
description = "ベクトルDBにデータ保存 / 曲調で検索する"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chromadb>=0.4.0",
    "colorama>=0.4.6",
    "librosa>=0.10.0",
    "numpy>=1.24.0",
    "ytmusicapi>=1.11.5",
    "streamlit>=1.31.0",
    "pandas>=2.0.0",
    "python-dotenv>=1.0.0",
]
